organized github auth routes and functions
