+++ src/routes/index.js
@@ -1,26 +1,30 @@
 const { findUser } = require("../../database/controllers/user.controller.js");
 const Repo = require("../../database/models/repo.model.js");
-const github_helpers = require("../helpers/github.js");
-const github_routes = require("./github.js");
+const github_helpers = require("../../providers/github/APICalls.js");
+const {
+  githubAuth,
+  githubAuthCallback,
+} = require("../../providers/github/auth.js");
 const gitlab_helpers = require("../helpers/gitlab.js");
 const gitlab_routes = require("./gitlab.js");
 
 /**
- * Redirects the user to the GitHub OAuth login page for authentication.
+ * Redirects the user to the OAuth login pages for authentication.
  * @param {*} req - object containing the client req details.
  * @param {*} res - object used to send a redirect response.
  */
-const login = (req, res) => {
-  const provider = req.query.provider;
 
-  if (provider === "github") {
-    github_helpers.authorizeApplication(res);
-  } else if (provider === "gitlab") {
-    gitlab_helpers.authorizeApplication(res);
-  } else {
-    res.status(400).send(`Unknown provider: ${provider}`);
-  }
-};
+// const login = (req, res) => {
+//   const provider = req.query.provider;
+
+//   if (provider === "github") {
+//     github_helpers.authorizeApplication(res);
+//   } else if (provider === "gitlab") {
+//     gitlab_helpers.authorizeApplication(res);
+//   } else {
+//     res.status(400).send(`Unknown provider: ${provider}`);
+//   }
+// };
 
 const reposToBadge = async (req, res) => {
   const selectedRepos = (await req.body.repos) || [];
@@ -97,11 +101,20 @@ const badgedRepos = async (req, res) => {
 };
 
 const setupRoutes = (app) => {
-  app.get("/api/login", login);
+  // logins
+  app.get("/api/auth/github", (req, res) => {
+    githubAuth(req, res);
+  });
+
+  //redirects
+  app.get("/api/callback/github", (req, res) => {
+    githubAuthCallback(req, res);
+  });
+
   app.get("/api/badgedRepos", badgedRepos);
   app.post("/api/repos-to-badge", reposToBadge);
 
-  github_routes.setupGitHubRoutes(app);
+  // github_routes.setupGitHubRoutes(app);
   gitlab_routes.setupGitLabRoutes(app);
 };
